requirements.txt
Django==1.8.4
django-reversion>=1.8.2,<1.8.6
html5lib>=0.90,!=0.9999,!=0.99999,<0.99999999
django-cms
djangocms-admin-style
django-filer
cmsplugin-filer
djangocms-text-ckeditor
djangocms-text-ckeditor
aldryn-newsblog
aldryn-people
aldryn-events
aldryn-faq
aldryn-jobs
aldryn-categories
aldryn-bootstrap3
aldryn-forms
aldryn-style
aldryn-locations
djangorestframework
BeautifulSoup
beautifulsoup4
whitenoise
gevent
gunicorn==19.4.5
